Special case handling for Tybal, lvl2 gazer and Stone golem on level 6 events caused by their deaths Door controller will trigger the correct trigger when used Fixed generation of Lava tiles so they now apply damage. Basic NPC damage calculation based on npc_power and cumulative armour score Fix to swapping objects in full container. Changed some data types

+++ UnityScripts/scripts/Loaders/ObjectLoaderInfo.cs
@@ -20,7 +20,7 @@ public class ObjectLoaderInfo {
 		public int y; //  13-15   3   "xpos"      Object X position (0-7)
 		//0004 quality / chain
 		public 		int quality;	//;     0- 5   6   "quality"   Quality
-		public long next; //    6-15   10  "next"      Index of next object in chain
+		public int next; //    6-15   10  "next"      Index of next object in chain
 
 		//0006 link / special
 		//     0- 5   6   "owner"     Owner / special
